+++ pyrgg/functions.py
@@ -29,49 +29,49 @@ def is_weighted(max_weight, min_weight, signed):
     return True
 
 
-def get_min_max_weight(weight_dic):
+def get_min_max_weight(weight_dict):
     """
     Get minimum and maximum weight values.
 
-    :param weight_dic: weight dictionary
-    :type weight_dic: dict
+    :param weight_dict: weight dictionary
+    :type weight_dict: dict
     :return: minimum and maximum weight values
     """
-    all_weights = [abs(w) for weights in weight_dic.values() for w in weights]
+    all_weights = [abs(w) for weights in weight_dict.values() for w in weights]
     return min(all_weights), max(all_weights)
 
 
-def is_signed(weight_dic): # pragma: no cover
+def is_signed(weight_dict): # pragma: no cover
     """
     Check if the graph is signed.
 
-    :param weight_dic: weight dictionary
-    :type weight_dic: dict
+    :param weight_dict: weight dictionary
+    :type weight_dict: dict
     :return: signed flag
     """
-    return any([any([w < 0 for w in weights]) for weights in weight_dic.values()])
+    return any([any([w < 0 for w in weights]) for weights in weight_dict.values()])
 
 
-def has_self_loop(edge_dic): # pragma: no cover
+def has_self_loop(edge_dict): # pragma: no cover
     """
     Check if the graph has self loops.
     
-    :param edge_dic: edge dictionary
-    :type edge_dic: dict
+    :param edge_dict: edge dictionary
+    :type edge_dict: dict
     :return: self looped flag
     """
-    return any([v in edges for v, edges in edge_dic.items()])
+    return any([v in edges for v, edges in edge_dict.items()])
 
 
-def is_multigraph(edge_dic):
+def is_multigraph(edge_dict):
     """
     Check if the graph is a multigraph.
 
-    :param edge_dic: edge dictionary
-    :type edge_dic: dict
+    :param edge_dict: edge dictionary
+    :type edge_dict: dict
     :return: multigraph flag
     """
-    return any([len(set(edges)) != len(edges) for edges in edge_dic.values()])
+    return any([len(set(edges)) != len(edges) for edges in edge_dict.values()])
 
 
 def get_precision(input_number):