+++ pyrgg/graph_gen.py
@@ -10,16 +10,16 @@ random_system = random
 
 
 def dimacs_maker(
-        edge_dic,
-        weight_dic,
+        edge_dict,
+        weight_dict,
         mdata):
     """
     Create output file and fill in.
 
-    :param edge_dic: dictionary containing edges data
-    :type edge_dic: dict
-    :param weight_dic: dictionary containing weights data
-    :type weight_dic: dict
+    :param edge_dict: dictionary containing edges data
+    :type edge_dict: dict
+    :param weight_dict: dictionary containing weights data
+    :type weight_dict: dict
     :param mdata: meta data
     :type mdata: dict
     :return: None
@@ -33,21 +33,21 @@ def dimacs_maker(
                 str(mdata['max_weight']),
                 str(mdata['min_weight'])))
         _write_separated_file(
-            buf, edge_dic, weight_dic, separator=' ', prefix='a',
+            buf, edge_dict, weight_dict, separator=' ', prefix='a',
         )
 
 
 def json_maker(
-        edge_dic,
-        weight_dic,
+        edge_dict,
+        weight_dict,
         mdata):
     """
     Create output file in json format.
 
-    :param edge_dic: dictionary containing edges data
-    :type edge_dic: dict
-    :param weight_dic: dictionary containing weights data
-    :type weight_dic: dict
+    :param edge_dict: dictionary containing edges data
+    :type edge_dict: dict
+    :param weight_dict: dictionary containing weights data
+    :type weight_dict: dict
     :param mdata: meta data
     :type mdata: dict
     :return: None
@@ -60,26 +60,26 @@ def json_maker(
             mdata['multigraph'],)
         _write_data_to_json(
             buf,
-            edge_dic,
-            weight_dic,
+            edge_dict,
+            weight_dict,
         )
 
 
-def _write_data_to_json(buf, edge_dic, weight_dic):
+def _write_data_to_json(buf, edge_dict, weight_dict):
     """Write data to json buffer.
 
     :param buf: output file object
     :type buf: file_object
-    :param edge_dic: dictionary containing edges data
-    :type edge_dic: dict
-    :param weight_dic: dictionary containing weights data
-    :type weight_dic: dict
+    :param edge_dict: dictionary containing edges data
+    :type edge_dict: dict
+    :param weight_dict: dictionary containing weights data
+    :type weight_dict: dict
     :return: None
     """
     buf.write('\n\t"graph": {\n')
-    _write_nodes_to_json(buf, edge_dic)
+    _write_nodes_to_json(buf, edge_dict)
     buf.write("\n\t\t],\n")
-    _write_edges_to_json(buf, edge_dic, weight_dic)
+    _write_edges_to_json(buf, edge_dict, weight_dict)
     buf.write("\n\t\t]\n\t}\n}")
 
 
@@ -108,20 +108,20 @@ def _write_properties_to_json(
     buf.write("},\n")
 
 
-def _write_nodes_to_json(buf, edge_dic):
+def _write_nodes_to_json(buf, edge_dict):
     """Write nodes to json.
 
     :param buf: output file object
     :type buf: file_object
-    :param edge_dic: dictionary containing edges data
-    :type edge_dic: dict
+    :param edge_dict: dictionary containing edges data
+    :type edge_dict: dict
     :return: None
     """
     first_line = True
     nodes = '\t\t"nodes":[\n'
     buf.write(nodes)
 
-    for key in edge_dic:
+    for key in edge_dict:
         nodes = ""
         if first_line:
             first_line = False
@@ -137,15 +137,15 @@ def _write_nodes_to_json(buf, edge_dic):
         buf.write(nodes)
 
 
-def _write_edges_to_json(buf, edge_dic, weight_dic):
+def _write_edges_to_json(buf, edge_dict, weight_dict):
     """Write edges to json.
 
     :param buf: output file object
     :type buf: file_object
-    :param edge_dic: dictionary containing edges data
-    :type edge_dic: dict
-    :param weight_dic: dictionary containing weights data
-    :type weight_dic: dict
+    :param edge_dict: dictionary containing edges data
+    :type edge_dict: dict
+    :param weight_dict: dictionary containing weights data
+    :type weight_dict: dict
     :return: None
     """
     edges = '\t\t"edges":[\n'
@@ -153,7 +153,7 @@ def _write_edges_to_json(buf, edge_dic, weight_dic):
     first_line = True
     buf.write(edges)
 
-    for key, edge_val in edge_dic.items():
+    for key, edge_val in edge_dict.items():
         for j, value in enumerate(edge_val):
             edges = ""
             if first_line:
@@ -169,59 +169,59 @@ def _write_edges_to_json(buf, edge_dic, weight_dic):
                 str(value),
                 ',\n\t\t\t',
                 '"weight": ',
-                str(weight_dic[key][j]),
+                str(weight_dict[key][j]),
                 '\n\t\t}'
             ])
             buf.write(edges)
 
 
 def csv_maker(
-        edge_dic,
-        weight_dic,
+        edge_dict,
+        weight_dict,
         mdata):
     """
     Create output file in csv format.
 
-    :param edge_dic: dictionary containing edges data
-    :type edge_dic: dict
-    :param weight_dic: dictionary containing weights data
-    :type weight_dic: dict
+    :param edge_dict: dictionary containing edges data
+    :type edge_dict: dict
+    :param weight_dict: dictionary containing weights data
+    :type weight_dict: dict
     :param mdata: meta data
     :type mdata: dict
     :return: None
     """
     with open(mdata['file_name'] + ".csv", "w") as buf:
-        _write_separated_file(buf, edge_dic, weight_dic, separator=',')
+        _write_separated_file(buf, edge_dict, weight_dict, separator=',')
 
 
 def tsv_maker(
-        edge_dic,
-        weight_dic,
+        edge_dict,
+        weight_dict,
         mdata):
     """
     Create output file in tsv format.
 
-    :param edge_dic: dictionary containing edges data
-    :type edge_dic: dict
-    :param weight_dic: dictionary containing weights data
-    :type weight_dic: dict
+    :param edge_dict: dictionary containing edges data
+    :type edge_dict: dict
+    :param weight_dict: dictionary containing weights data
+    :type weight_dict: dict
     :param mdata: meta data
     :type mdata: dict
     :return: None
     """
     with open(mdata['file_name'] + ".tsv", "w") as buf:
-        _write_separated_file(buf, edge_dic, weight_dic, separator='\t')
+        _write_separated_file(buf, edge_dict, weight_dict, separator='\t')
 
 
-def _write_separated_file(buf, edge_dic, weight_dic, separator, prefix=''):
+def _write_separated_file(buf, edge_dict, weight_dict, separator, prefix=''):
     r"""Write data to buffer separated with ``separator``.
 
     :param buf: output file object
     :type buf: file_object
-    :param edge_dic: dictionary containing edges data
-    :type edge_dic: dict
-    :param weight_dic: dictionary containing weights data
-    :type weight_dic: dict
+    :param edge_dict: dictionary containing edges data
+    :type edge_dict: dict
+    :param weight_dict: dictionary containing weights data
+    :type weight_dict: dict
     :param separator: separator in a separated file, like ',', '\t', ' ', etc.
     :type separator: str
     :param prefix: prefix to be added in front of each line
@@ -231,48 +231,48 @@ def _write_separated_file(buf, edge_dic, weight_dic, separator, prefix=''):
     dummy_prefix = object()
     prefix = prefix or dummy_prefix
 
-    for key, edge_val in edge_dic.items():
+    for key, edge_val in edge_dict.items():
         for j, value in enumerate(edge_val):
             elements = [
                 prefix,
                 str(key),
                 str(value),
-                str(weight_dic[key][j]) + "\n"
+                str(weight_dict[key][j]) + "\n"
             ]
             string = separator.join(x for x in elements if x != dummy_prefix)
             buf.write(string)
 
 
 def wel_maker(
-        edge_dic,
-        weight_dic,
+        edge_dict,
+        weight_dict,
         mdata):
     """
     Create output file in wel format.
 
-    :param edge_dic: dictionary containing edges data
-    :type edge_dic: dict
-    :param weight_dic: dictionary containing weights data
-    :type weight_dic: dict
+    :param edge_dict: dictionary containing edges data
+    :type edge_dict: dict
+    :param weight_dict: dictionary containing weights data
+    :type weight_dict: dict
     :param mdata: meta data
     :type mdata: dict
     :return: None
     """
     with open(mdata['file_name'] + ".wel", "w") as buf:
-        _write_separated_file(buf, edge_dic, weight_dic, separator=' ')
+        _write_separated_file(buf, edge_dict, weight_dict, separator=' ')
 
 
 def mtx_maker(
-        edge_dic,
-        weight_dic,
+        edge_dict,
+        weight_dict,
         mdata):
     """
     Create output file in Matrix Market format.
 
-    :param edge_dic: dictionary containing edges data
-    :type edge_dic: dict
-    :param weight_dic: dictionary containing weights data
-    :type weight_dic: dict
+    :param edge_dict: dictionary containing edges data
+    :type edge_dict: dict
+    :param weight_dict: dictionary containing weights data
+    :type weight_dict: dict
     :param mdata: meta data
     :type mdata: dict
     :return: None
@@ -283,140 +283,140 @@ def mtx_maker(
         buf.write(
             "{0}    {0}    {1}\n".format(str(mdata['vertices_number']), str(mdata['edge_number']))
         )
-        for key, edge_vals in edge_dic.items():
+        for key, edge_vals in edge_dict.items():
             for j, value in enumerate(edge_vals):
                 shift1 = (max_edges_length - len(str(key))) + 4
                 shift2 = (max_edges_length - len(str(value))) + 4
                 buf.write(str(key) + shift1 * " " + str(value) + shift2 * " " +
-                          str(weight_dic[key][j]) + "\n")
+                          str(weight_dict[key][j]) + "\n")
 
 
 def lp_maker(
-        edge_dic,
-        weight_dic,
+        edge_dict,
+        weight_dict,
         mdata):
     """
     Create output file in ASP format.
 
-    :param edge_dic: dictionary containing edges data
-    :type edge_dic: dict
-    :param weight_dic: dictionary containing weights data
-    :type weight_dic: dict
+    :param edge_dict: dictionary containing edges data
+    :type edge_dict: dict
+    :param weight_dict: dictionary containing weights data
+    :type weight_dict: dict
     :param mdata: meta data
     :type mdata: dict
     :return: None
     """
     with open(mdata['file_name'] + ".lp", "w") as buf:
-        for key in edge_dic:
+        for key in edge_dict:
             buf.write('node(' + str(key) + ").\n")
-        for key, edge_val in edge_dic.items():
+        for key, edge_val in edge_dict.items():
             for j, value in enumerate(edge_val):
                 buf.write('edge(' + str(key) + "," + str(value) +
-                          "," + str(weight_dic[key][j]) + ").\n")
+                          "," + str(weight_dict[key][j]) + ").\n")
 
 
 def tgf_maker(
-        edge_dic,
-        weight_dic,
+        edge_dict,
+        weight_dict,
         mdata):
     """
     Create output file in Trivial Graph Format (TGF).
 
-    :param edge_dic: dictionary containing edges data
-    :type edge_dic: dict
-    :param weight_dic: dictionary containing weights data
-    :type weight_dic: dict
+    :param edge_dict: dictionary containing edges data
+    :type edge_dict: dict
+    :param weight_dict: dictionary containing weights data
+    :type weight_dict: dict
     :param mdata: meta data
     :type mdata: dict
     :return: None
     """
     with open(mdata['file_name'] + ".tgf", "w") as buf:
-        for key in edge_dic:
+        for key in edge_dict:
             buf.write(str(key) + "\n")
         buf.write("#\n")
-        _write_separated_file(buf, edge_dic, weight_dic, separator=' ')
+        _write_separated_file(buf, edge_dict, weight_dict, separator=' ')
 
 
 def gl_maker(
-        edge_dic,
-        weight_dic,
+        edge_dict,
+        weight_dict,
         mdata):
     """
     Create output file in Graph Line(GL).
 
-    :param edge_dic: dictionary containing edges data
-    :type edge_dic: dict
-    :param weight_dic: dictionary containing weights data
-    :type weight_dic: dict
+    :param edge_dict: dictionary containing edges data
+    :type edge_dict: dict
+    :param weight_dict: dictionary containing weights data
+    :type weight_dict: dict
     :param mdata: meta data
     :type mdata: dict
     :return: None
     """
     with open(mdata['file_name'] + ".gl", "w") as buf:
-        for key, edge_val in edge_dic.items():
+        for key, edge_val in edge_dict.items():
             line_data = str(key)
             write_flag = False
             for j, value in enumerate(edge_val):
                 write_flag = True
-                line_data += " " + str(value) + ":" + str(weight_dic[key][j])
+                line_data += " " + str(value) + ":" + str(weight_dict[key][j])
             if write_flag:
                 buf.write(line_data + "\n")
 
 
 def dl_maker(
-        edge_dic,
-        weight_dic,
+        edge_dict,
+        weight_dict,
         mdata):
     """
     Create output file in UCINET DL Format.
 
-    :param edge_dic: dictionary containing edges data
-    :type edge_dic: dict
-    :param weight_dic: dictionary containing weights data
-    :type weight_dic: dict
+    :param edge_dict: dictionary containing edges data
+    :type edge_dict: dict
+    :param weight_dict: dictionary containing weights data
+    :type weight_dict: dict
     :param mdata: meta data
     :type mdata: dict
     :return: None
     """
     with open(mdata['file_name'] + ".dl", "w") as buf:
         buf.write("dl\nformat=edgelist1\nn=" + str(mdata['vertices_number']) + "\ndata:\n")
-        _write_separated_file(buf, edge_dic, weight_dic, separator=' ')
+        _write_separated_file(buf, edge_dict, weight_dict, separator=' ')
 
 
 def gdf_maker(
-        edge_dic,
-        weight_dic,
+        edge_dict,
+        weight_dict,
         mdata):
     """
     Create output file in GDF Format.
 
-    :param edge_dic: dictionary containing edges data
-    :type edge_dic: dict
-    :param weight_dic: dictionary containing weights data
-    :type weight_dic: dict
+    :param edge_dict: dictionary containing edges data
+    :type edge_dict: dict
+    :param weight_dict: dictionary containing weights data
+    :type weight_dict: dict
     :param mdata: meta data
     :type mdata: dict
     :return: None
     """
     with open(mdata['file_name'] + ".gdf", "w") as buf:
         buf.write("nodedef>name VARCHAR,label VARCHAR\n")
-        for key in edge_dic:
+        for key in edge_dict:
             buf.write(str(key) + "," + "Node{0}".format(str(key)) + "\n")
         buf.write("edgedef>node1 VARCHAR,node2 VARCHAR,weight DOUBLE\n")
-        _write_separated_file(buf, edge_dic, weight_dic, separator=',')
+        _write_separated_file(buf, edge_dict, weight_dict, separator=',')
 
 
 def gml_maker(
-        edge_dic,
-        weight_dic,
+        edge_dict,
+        weight_dict,
         mdata):
     """
     Create output file in GML Format.
 
-    :param edge_dic: dictionary containing edges data
-    :type edge_dic: dict
-    :param weight_dic: dictionary containing weights data
-    :type weight_dic: dict
+    :param edge_dict: dictionary containing edges data
+    :type edge_dict: dict
+    :param weight_dict: dictionary containing weights data
+    :type weight_dict: dict
     :param mdata: meta data
     :type mdata: dict
     :return: None
@@ -426,7 +426,7 @@ def gml_maker(
 
     with open(mdata['file_name'] + ".gml", "w") as buf:
         buf.write(header)
-        for key in edge_dic:
+        for key in edge_dict:
             buf.write(
                 "  node\n  [\n   id " +
                 str(key) +
@@ -434,7 +434,7 @@ def gml_maker(
                 '   label "Node {0}"\n'.format(
                     str(key)) +
                 "  ]\n")
-        for key, edge_vals in edge_dic.items():
+        for key, edge_vals in edge_dict.items():
             for j, value in enumerate(edge_vals):
                 buf.write("  edge\n  [\n   source " +
                           str(key) +
@@ -443,23 +443,23 @@ def gml_maker(
                           str(value) +
                           "\n" +
                           "   value " +
-                          str(weight_dic[key][j]) +
+                          str(weight_dict[key][j]) +
                           "\n" +
                           "  ]\n")
         buf.write("]")
 
 
 def gexf_maker(
-        edge_dic,
-        weight_dic,
+        edge_dict,
+        weight_dict,
         mdata):
     """
     Create output file in GEXF Format.
 
-    :param edge_dic: dictionary containing edges data
-    :type edge_dic: dict
-    :param weight_dic: dictionary containing weights data
-    :type weight_dic: dict
+    :param edge_dict: dictionary containing edges data
+    :type edge_dict: dict
+    :param weight_dict: dictionary containing weights data
+    :type weight_dict: dict
     :param mdata: meta data
     :type mdata: dict
     :return: None
@@ -483,7 +483,7 @@ def gexf_maker(
             " " * 4 + '<graph defaultedgetype="' + defaultedgetype + '">\n'
         )
         buf.write(" " * 8 + "<nodes>\n")
-        for key in edge_dic:
+        for key in edge_dict:
             buf.write(
                 " " * 12 +
                 '<node id="' +
@@ -493,7 +493,7 @@ def gexf_maker(
         buf.write(" " * 8 + "</nodes>\n")
         buf.write(" " * 8 + "<edges>\n")
         edge_id = 1
-        for key, edge_vals in edge_dic.items():
+        for key, edge_vals in edge_dict.items():
             for j, value in enumerate(edge_vals):
                 buf.write(
                     " " * 12 +
@@ -504,7 +504,7 @@ def gexf_maker(
                     ' target="' +
                     str(value) + '"' +
                     ' weight="{0}" />'.format(
-                        str(weight_dic[key][j])) + "\n")
+                        str(weight_dict[key][j])) + "\n")
                 edge_id += 1
         buf.write(" " * 8 + "</edges>\n")
         buf.write(" " * 4 + "</graph>\n")
@@ -512,16 +512,16 @@ def gexf_maker(
 
 
 def dot_maker(
-        edge_dic,
-        weight_dic,
+        edge_dict,
+        weight_dict,
         mdata):
     """
     Create output file in Dot Format.
 
-    :param edge_dic: dictionary containing edges data
-    :type edge_dic: dict
-    :param weight_dic: dictionary containing weights data
-    :type weight_dic: dict
+    :param edge_dict: dictionary containing edges data
+    :type edge_dict: dict
+    :param weight_dict: dictionary containing weights data
+    :type weight_dict: dict
     :param mdata: meta data
     :type mdata: dict
     :return: None
@@ -536,7 +536,7 @@ def dot_maker(
 
     with open(mdata['file_name'] + ".gv", "w") as buf:
         buf.write(header + " {")
-        for key, edge_val in edge_dic.items():
+        for key, edge_val in edge_dict.items():
             for j, value in enumerate(edge_val):
                 buf.write(
                     "\n" +
@@ -546,6 +546,6 @@ def dot_maker(
                     " " +
                     str(value) +
                     " [weight={}]".format(
-                        weight_dic[key][j]) +
+                        weight_dict[key][j]) +
                     ";")
         buf.write("\n}")