+++ pyrgg/test.py
@@ -331,7 +331,7 @@ a 100 40 11
 >>> dimacs_maker({})
 Traceback (most recent call last):
         ...
-TypeError: dimacs_maker() missing 2 required positional arguments: 'weight_dic' and 'mdata'
+TypeError: dimacs_maker() missing 2 required positional arguments: 'weight_dict' and 'mdata'
 >>> random.seed(2)
 >>> engine.gen_using(json_maker, 'testfile', {'min_weight':0, 'max_weight':200, 'vertices':10, 'min_edges':0, 'max_edges':2, 'sign':1, 'direct':1, 'self_loop':1, 'multigraph':0})
 7
@@ -842,7 +842,7 @@ edgedef>node1 VARCHAR,node2 VARCHAR,weight DOUBLE
 >>> csv_maker({})
 Traceback (most recent call last):
         ...
-TypeError: csv_maker() missing 2 required positional arguments: 'weight_dic' and 'mdata'
+TypeError: csv_maker() missing 2 required positional arguments: 'weight_dict' and 'mdata'
 >>> random.seed(2)
 >>> engine.gen_using(wel_maker, 'testfile', {'min_weight':0, 'max_weight':200, 'vertices':10, 'min_edges':0, 'max_edges':2, 'sign':1, 'direct':1, 'self_loop':1, 'multigraph':0})
 7